requirements.txt
Django==5.2.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pathlib==1.0.1
celery==5.3.6
redis==5.0.1
serpapi==0.1.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0 